+++ src/components/App.jsx
@@ -1,36 +1,7 @@
-import { useState } from 'react';
-import { useDispatch } from 'react-redux';
-
 import RestaurantsContainer from './RestaurantsContainer';
-import RestaurantForm from './RestaurantForm';
-
-import { addNewRestaurant } from '../store/actions';
-import { Restaurant } from '../@types';
+import RestaurantsFormContainer from './RestaurantsFormContainer';
 
 export default function App() {
-  const [inputValues, setInputValues] = useState({
-    name: '',
-    category: '',
-    location: '',
-  });
-
-  const dispatch = useDispatch();
-
-  const updateRestaurantForm = (e) => {
-    setInputValues((prev) => ({
-      ...prev,
-      [e.target.name]: e.target.value,
-    }));
-  };
-
-  const submitNewRestaurantForm = (e) => {
-    e.preventDefault();
-
-    const newRestaurant = new Restaurant(inputValues);
-
-    dispatch(addNewRestaurant(newRestaurant));
-  };
-
   return (
     <div>
       <header>
@@ -38,11 +9,7 @@ export default function App() {
       </header>
 
       <RestaurantsContainer />
-      <RestaurantForm
-        inputValues={inputValues}
-        onUpdateRestaurantForm={updateRestaurantForm}
-        onAddNewRestaurant={submitNewRestaurantForm}
-      />
+      <RestaurantsFormContainer />
     </div>
   );
 }
